Bump denedencies, 'eslint' version, and run 'eslint --fix'

+++ packages/contracts/src/arbitrum/ArbSys.sol
@@ -10,7 +10,7 @@ interface ArbSys {
     * @notice Get Arbitrum block number (distinct from L1 block number; Arbitrum genesis block has block number 0)
     * @return block number as int
      */
-    function arbBlockNumber() external view returns (uint);
+    function arbBlockNumber() external view returns (uint256);
 
     /**
      * @notice get the caller's amount of available storage gas
